requirements.txt
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
dj-database-url==0.5.0
Django==4.0.3
django-contrib-postgres==0.0.1
django-cors-headers==3.11.0
django-postgres-extensions==0.9.3
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.11.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==6.0.0
